backend/requirements.txt
django==2.2
django-configurations==2.1
django-dotenv==1.4.2
dj-database-url==0.5.0
psycopg2-binary==2.7.7

# Production
gunicorn==19.9.0